Add mainclass, rename template package to reflect other tools structure
